Resolved conflicts:
- TipoEventoNotificacao: renumbered CAP/CAR enum values (32-36) after master's entries
- Permissao: appended CAP/CAR permissions after master's Helpdesk entries
- ServiceCollectionExtensions: added Helpdesk + Financeiro registrations
- EasyStockDbContext: kept both Lancamento and CAP/CAR DbSets
- _Sidebar.cshtml: added Financeiro/CAR/CAP nav items to Cadastros section (es-icon style)
- ModelSnapshot: accepted PR version (includes CAP/CAR entity configs)
- .knowledge: accepted master version

Build fixes:
- Admin Edit.cshtml.cs: added missing CarregarDuplicandoAsync method, fixed 'origem' reference
- FornecedorControllerTests: added missing GerarContaPagarDePedidoFornecedorUseCase parameter
- IntegrationTests.csproj: pinned EF Core Relational 9.0.4 to resolve version conflict

+++ EasyStock.Application/UseCases/BuscarEstoqueInteligente/BuscarEstoqueInteligenteUseCase.cs
@@ -29,7 +29,7 @@ namespace EasyStock.Application.UseCases.BuscarEstoqueInteligente
         string ChaveExibicao,
         int Score,
         string? Sku = null,
-        int? QuantidadeAtual = null,
+        decimal? QuantidadeAtual = null,
         string? Status = null,
         string? FornecedorNome = null,
         string? Loja = null);
